Refactor: Move onboarding routes out of UsersController
